# EdgeTPU optimizations