The EdgeTPU compiler seems to have issues with the transpose operation. This pull request removes this op by applying it to the constant weights of the subsequent add and mul ops. As a result, more operations can be mapped onto the EdgeTPU. Reducing the imgsz parameter for the export script to 160 will lead to a model which is mapped entirely.