@zldrobit 2022-03-05
@paradigmn I could confirm that using this PR the exported EdgeTPU model with 160x160 input size contains only one Edge TPU op.
![image](https://user-images.githubusercontent.com/3896992/156864899-d3432a79-e1b6-4953-9da5-e1267a00b051.png)
It generates a more compact Edge TPU model as well as with 640x640 input size.

<details>
<summary>Original Edge TPU model (above) and Optimized Edge TPU model (below)</summary>

![image](https://user-images.githubusercontent.com/3896992/156864970-5cf281fb-ff02-44a5-9d8d-19d063f259b9.png)
![image](https://user-images.githubusercontent.com/3896992/156864982-168dbc14-8ce5-4a4e-913c-c1edbc4d57f4.png)

</details>
The inference time of the 640x640 model is reduced from 270ms to 180ms.
I ran validations for SavedModel, GraphDef, TFLite and TF.js, and they all passed.
Though removing transpose ops improve the compatibility with OpenGL delegate, `Add` op with HWC tensor still hinders the model from running on GPU with TFLite OpenGL delegate.

IMHO, the code looks good and we should merge this PR.