@zldrobit 2022-03-05
@glenn-jocher Maybe we could further move the transpose op from `Detect()/TFDetect()` to training code for an easier understanding and simplicity.