@zldrobit 2022-03-06
>@zldrobit glad to hear the code works as desired. Is the issue with the GPU delegate caused by my PR or is it generally not working? I am not developing on Android, so I can't do any test in this regard.

@paradigmn The problem on Android can be solved by https://github.com/ultralytics/yolov5/pull/1127#issuecomment-1056038455, and it is not caused by this PR. Plz dont' worry about the situation on Android.

>Is there a reason the "experimental_new_quantizer" is disabled in the export script? This option should improved model speed by removing redundant quantize and dequantize ops.

The "experimental_new_quantizer" is disabled because it introduces huge quantization error when converting YOLOv5 models to TFLite int8 models.

>Another micro optimization is stumbled upon: you can further avoid a redundant sub and mul op per grid if you perform them at compile time (along with the transpose op).
>```
>grid = tf.transpose(self.grid[i], [0, 2, 1, 3]) - 0.5
>anchor_grid = tf.transpose(self.anchor_grid[i], [0, 2, 1, 3]) * 4
>xy = (y[..., 0:2] * 2 + grid) * self.stride[i]  # xy
>wh = y[..., 2:4] ** 2 * anchor_grid
>```
>Should I commit this here as well or create another PR, once this one is merged?

I suggest you keep this change in another PR, so this PR could be merged as soon as possible. You don't have to wait this PR to be merged before creating another PR. By the same time, you could create another PR based on the commits of this PR. In general, we could have parallel PRs.