@paradigmn 2022-03-06
>The "experimental_new_quantizer" is disabled because it introduces huge quantization error when converting YOLOv5 models to TFLite int8 models.
>
@zldrobit are you certain this is still an issue with the current Tensorflow version? I made a quick test with the val.py script and TF2.8. It actually shows an improvement in accuracy with the flag enabled for YOLO5s@640.
|            | OFF   | ON    |
|------------|-------|-------|
| mAP@.5     | 0.521 | 0.536 |
| mAP@.5:.95 | 0.304 | 0.319 |
>Considering the purpose of this PR is to remove transpose ops I suggest you keep this change in another PR, so this PR could be merged as soon as possible. You don't have to wait this PR to be merged before creating another PR. By the same time, you could create another PR based on the commits of this PR. In general, we could have parallel PRs.
>
Alright, I'll set up a new PR.