@zldrobit 2022-03-07
> @zldrobit are you certain this is still an issue with the current Tensorflow version? I made a quick test with the val.py script and TF2.8. It actually shows an improvement in accuracy with the flag enabled for YOLO5s@640.

@paradigmn Thanks for pointing out this. I re-exported the YOLOv5s@640 and YOLOv5s@320 with TF2.8, and ran val.py. I could confirm that in both cases the accuracies do not drop after turning on the `experimental_new_quantizer` option. Enabling this option removes most of `Quantize` ops.
<details>
<summary>For example, 11 `Quantize` ops are removed near the output tensor (before Edge TPU compilation): disable (upper) and enable (below):
</summary>

![image](https://user-images.githubusercontent.com/3896992/156995104-55c6446c-48dc-4c51-bca6-b708dfa69eab.png)
![image](https://user-images.githubusercontent.com/3896992/156994620-ab502bcd-ad69-4344-bf86-7e9092143c78.png)
</details>

On Edge TPU, the inference time of YOLOv5s@640 with `experimental_new_quantizer` is reduced to 145ms.

>Alright, I'll set up a new PR.

Forget about what I said, I didn't conceive there would be more and more changes for Edge TPU. Considering we are talking about optimizations on EdgeTPU, what about changing the title of this PR? So we could gather all optimizations for Edge TPU in one thread.