@glenn-jocher 2022-03-11
> @glenn-jocher Maybe we could further move the transpose op from `Detect()/TFDetect()` to training code for an easier understanding and simplicity.

The permutation here is actually used in training and inference. It might be restructured elsewhere like NMS and loss function, but this is a big change, let's consider this later so we can get this PR merged now.

https://github.com/ultralytics/yolov5/blob/b94b59e199047aa8bf2cdd4401ae9f5f42b929e6/models/yolo.py#L54