@glenn-jocher 2022-03-11
@paradigmn @zldrobit CPU benchmark results on Colab++ high-mem. Couldn't profile Edge TPU but everything else seems fine.

YOLOv5 🚀 v6.1-27-gb94b59e torch 1.10.0+cu111 CPU

```python
!git clone https://github.com/ultralytics/yolov5  # clone
!git clone https://github.com/paradigmn/yolov5 yolov5-pr -b edgetpu-optimizations  # clone
%pip install -r yolov5/requirements.txt coremltools onnx onnx-simplifier onnxruntime openvino-dev tensorflow-cpu

!python yolov5/utils/benchmarks.py
!python yolov5-pr/utils/benchmarks.py
```

### master

```
Benchmarks complete (476.14s)
                   Format  mAP@0.5:0.95  Inference time (ms)
0                 PyTorch      0.462296           113.828156
1             TorchScript      0.462296           162.735077
2                    ONNX      0.462296            69.062775
3                OpenVINO      0.462296            73.463997
4                TensorRT           NaN                  NaN
5                  CoreML           NaN                  NaN
6   TensorFlow SavedModel      0.462296           127.020458
7     TensorFlow GraphDef      0.462296           126.529440
8         TensorFlow Lite      0.462334           239.746148
9     TensorFlow Edge TPU           NaN                  NaN
10          TensorFlow.js           NaN                  NaN
```

### PR

```
Benchmarks complete (460.14s)
                   Format  mAP@0.5:0.95  Inference time (ms)
0                 PyTorch      0.462296           120.689692
1             TorchScript      0.462296           147.933541
2                    ONNX      0.462296            67.923378
3                OpenVINO      0.462296            71.461299
4                TensorRT           NaN                  NaN
5                  CoreML           NaN                  NaN
6   TensorFlow SavedModel      0.462296           125.076026
7     TensorFlow GraphDef      0.462296           125.489706
8         TensorFlow Lite      0.462334           238.644088
9     TensorFlow Edge TPU           NaN                  NaN
10          TensorFlow.js           NaN                  NaN
```